Issue #3891: reorganized xpath package inputs
