+++ src/test/java/com/puppycrawl/tools/checkstyle/xpath/RootNodeTest.java
@@ -20,7 +20,6 @@
 package com.puppycrawl.tools.checkstyle.xpath;
 
 import static com.puppycrawl.tools.checkstyle.internal.XpathUtil.getXpathItems;
-
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -32,6 +31,7 @@ import java.util.List;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.puppycrawl.tools.checkstyle.AbstractPathTestSupport;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.TokenTypes;
 import com.puppycrawl.tools.checkstyle.internal.TestUtils;
@@ -40,14 +40,18 @@ import net.sf.saxon.om.Item;
 import net.sf.saxon.om.NamespaceBinding;
 import net.sf.saxon.tree.iter.EmptyIterator;
 
-public class RootNodeTest {
+public class RootNodeTest extends AbstractPathTestSupport {
 
     private static RootNode rootNode;
 
+    @Override
+    protected String getPackageLocation() {
+        return "com/puppycrawl/tools/checkstyle/xpath/xpathmapper";
+    }
+
     @Before
     public void init() throws Exception {
-        final File file = new File("src/test/resources/com/puppycrawl/tools/"
-                + "checkstyle/xpath/InputXpathMapperAst.java");
+        final File file = new File(getPath("InputXpathMapperAst.java"));
         final DetailAST rootAst = TestUtils.parseFile(file);
         rootNode = new RootNode(rootAst);
     }
