--- src/test/resources/com/puppycrawl/tools/checkstyle/xpath/InputXpathAnnotation.java
@@ -1,6 +1,0 @@
-package com.puppycrawl.tools.checkstyle.xpath;
-
-@SuppressWarnings("test")
-public class InputXpathAnnotation {
-
-}
