--- src/test/resources/com/puppycrawl/tools/checkstyle/xpath/InputXpathEnum.java
@@ -1,7 +1,0 @@
-package com.puppycrawl.tools.checkstyle.xpath;
-
-public enum InputXpathEnum {
-    ONE,
-    TWO,
-    THREE
-}
