+++ src/test/resources/com/puppycrawl/tools/checkstyle/xpath/xpathmapper/InputXpathMapperAnnotation.java
@@ -1,0 +1,6 @@
+package com.puppycrawl.tools.checkstyle.xpath.xpathmapper;
+
+@SuppressWarnings("test")
+public class InputXpathMapperAnnotation {
+
+}
