+++ src/test/resources/com/puppycrawl/tools/checkstyle/xpath/xpathmapper/InputXpathMapperEnum.java
@@ -1,0 +1,7 @@
+package com.puppycrawl.tools.checkstyle.xpath.xpathmapper;
+
+public enum InputXpathMapperEnum {
+    ONE,
+    TWO,
+    THREE
+}
